Move analysis to separate module
- Create a new module to hold the Submission class
- Move some common functionality into Submission class methods
- Move a bunch more common functions into utils
- Create a new analysis module to hold the functions for generating and printing analyses
- Split the quarter count logic out of the get_analysis function
- Add an option to the csv_tools script to run analysis on data before outputting it

TODO: Rework some of the script logic, probably pull actually useful commands all into one place, clean up arguments, etc.
TODO: Put submission choosing logic somewhere more useful
TODO: Put all the functions related to actually generating the seeds more useful, too
TODO: Clean up terminology for seed/order/slot???
