+++ seeding/utils.py
@@ -1,11 +1,18 @@
+import re
 import itertools
 import math
+import unicodedata
 
 __all__ = (
     "chunk",
     "get_bracket_size",
+    "get_canonical_artist",
+    "get_canonical_submitter",
     "get_distance",
+    "get_number_of_rounds",
+    "get_shuffled_range",
     "has_byes",
+    "insert_byes",
     "invert_list",
 )
 
@@ -27,6 +34,70 @@ def chunk(lst, n):
     return iter(lambda: tuple(itertools.islice(it, n)), ())
 
 
+def get_bracket_size(n):
+    """
+    Given a total number of submissions, returns the best fit bracket size
+    """
+
+    if 96 <= n < 128:
+        return 96
+    elif 48 <= n < 64:
+        return 48
+    else:
+        return 2 ** math.floor(math.log2(n))
+
+
+def get_canonical_artist(artist):
+    """
+    Converts artist name string to canonical version, for comparison.
+
+    This performs the following operations, in order:
+        - lowercases the input string
+        - replaces double hyphens with a space
+        - strips diacritics and many other miscellaneous marks (like ``&``)
+        - collapses multiple spaces in a row to a single space
+        - strips leading "The"
+        - drops featured artists from the end of the artist name by looking for:
+            - ``ft``, ``feat``, or ``featuring``
+            - optional period after the two abbreviations
+            - optional parentheses around the whole thing
+            - must have something following the "featuring" introduction, strips
+              to the end of the artist name
+
+    :param artist: String artist name
+    :returns: Canonical artist name, suitable for comparison
+    """
+
+    def should_keep(c):
+        return unicodedata.category(c)[0] in {"L", "N", "S", "Z"}
+
+    return re.sub(
+        r" (\()?(?:f(?:ea)?t\.?|featuring) .+(?(1)\))$",
+        "",
+        re.sub(
+            r"^the ",
+            "",
+            re.sub(
+                r"\s+",
+                " ",
+                "".join(
+                    filter(
+                        should_keep,
+                        unicodedata.normalize(
+                            "NFKD",
+                            artist.lower().replace("--", " "),
+                        ),
+                    ),
+                ),
+            ),
+        ),
+    )
+
+
+def get_canonical_submitter(submitter):
+    return submitter.lower()
+
+
 def get_distance(i, j):
     """
     Calculates the number of rounds before two songs would meet in a match.
@@ -43,17 +114,23 @@ def get_distance(i, j):
     return math.floor(math.log2(i ^ j))
 
 
-def get_bracket_size(n):
+def get_number_of_rounds(n):
     """
-    Given a total number of submissions, returns the best fit bracket size
+    Returns the total number of rounds for a bracket of size n.
     """
 
-    if 96 <= n < 128:
-        return 96
-    elif 48 <= n < 64:
-        return 48
-    else:
-        return 2 ** math.floor(math.log2(n))
+    return get_distance(0, n - 1) + 1
+
+
+def get_shuffled_range(n):
+    """
+    Utility function to generate a random ordering of the integers in [0, n)
+
+    :param n: The length of the returned list
+    :returns: A list of the numbers from 0 to n-1, inclusive, shuffled
+    """
+
+    return random.sample(list(range(n)), k=n)
 
 
 def has_byes(n):
@@ -62,6 +139,29 @@ def has_byes(n):
     """
 
     return n in {48, 96}
+
+
+def insert_byes(seeds):
+    """
+    If byes are required, inserts dummy seeds for the bye rows.
+
+    If any changes are required, returns a new copy of the seeds list with the
+    bye seeds interleaved.
+    """
+
+    n = len(seeds)
+
+    if has_byes(n):
+        # Every fourth submission should be a bye
+        # The bye submissions are all at the end, so adding n to a 0-based index
+        # will give a bye submission index
+        seeds = list(
+            itertools.chain.from_iterable(
+                trio + (i + n,) for i, trio in enumerate(chunk(seeds, 3))
+            )
+        )
+
+    return seeds
 
 
 def invert_list(lst):
